Renamed the mouse grids file
